Move jsonToFromat to where it's used
